# chore(perf): fix catch micro perf test