--- a/perf/micro/immediate-scheduler/operators/catch.js
+++ b/perf/micro/immediate-scheduler/operators/catch.js
@@ -3,8 +3,10 @@ var RxNew = require("../../../../index");
 
 module.exports = function (suite) {
 
-  var oldCatchWithImmediateScheduler = RxOld.Observable.throw(new Error('error'), RxOld.Scheduler.immediate).catch(RxOld.Observable.of(25, RxOld.Scheduler.immediate));
-  var newCatchWithImmediateScheduler = RxNew.Observable.throw(new Error('error')).catch(RxNew.Observable.of(25));
+  var oldCatchWithImmediateScheduler = RxOld.Observable.throw(new Error('error'), RxOld.Scheduler.immediate)
+    .catch(function(){ return RxOld.Observable.of(25, RxOld.Scheduler.immediate); });
+  var newCatchWithImmediateScheduler = RxNew.Observable.throw(new Error('error'))
+    .catch(function() { return RxNew.Observable.of(25); });
 
   return suite
     .add('old catch with immediate scheduler', function () {