The new method isn't polymorphic, and can only be called with a function

```sh
old catch with immediate scheduler x 235,384 ops/sec ±0.64% (94 runs sampled)
new catch with immediate scheduler x 1,621,709 ops/sec ±0.56% (96 runs sampled)
	588.96% faster than Rx2

old catch with current thread scheduler x 217,551 ops/sec ±1.02% (94 runs sampled)
new catch with current thread scheduler x 637,311 ops/sec ±0.54% (93 runs sampled)
	192.95% faster than Rx2
```

closes #283